strauss: rely on strace--tips test instead of forcibly trimming lines
The regression introduced in commit v5.19~99 "strauss: tips whitespace
and phrasing cleanups" (and subsequently fixed in v6.2-6-g5b4f05f42
"strauss: add missing line separators") highlights the issue
with the current approach with respect to tips formatting: it masks
the issues related to overlong lines in favor to the pretty formatting;
however, the presence of a (working) strace--tips test, that checks the
correctness of the tips rendering, makes this trimming superfluous.
Drop it and rely on the testing with respect to the guaranteeing
the proper wrapping of (overly) long tip lines.

* src/strauss.c (print_totd): Do not provide precision to the format
specification that prints tip line.

References: v6.2-6-g5b4f05f42 "strauss: add missing line separators"