exec_or_die: remove incomprehensible setreuid invocation
Since the very first commit setreuid is invoked in geteuid() != 0
case without any plausible explanation.  It makes little sense to
install strace as a setuid-nonroot executable in the first place,
and even in that case it is unclear why one would like to drop the
effective nonzero uid before the execve.

* src/strace.c (exec_or_die) <!username && geteuid() != 0>: Remove.