strace.spec: try to comply with the new Fedora package license policy
New Fedora policies[1][2] require the usage of the SPDX identifiers
with pretty specific syntax[3], and [4] requires refraining from any
"license analysis" and listing all the licenses verbatim (which is especially
fun with all the bundled sources), so here we are.  The %changelog entry
mention is added per [5].

[1] https://fedoraproject.org/wiki/Changes/SPDX_Licenses_Phase_2
[2] https://docs.fedoraproject.org/en-US/legal/update-existing-packages/
[3] https://docs.fedoraproject.org/en-US/legal/license-field/#_conjunctive_and_licensing
[4] https://docs.fedoraproject.org/en-US/legal/license-field/#_no_effective_license_analysis
[5] https://docs.fedoraproject.org/en-US/legal/update-existing-packages/#_changelog_entry

* strace.spec.in [0%{?fedora} >= 35 || 0%{?centos} >= 9 || 0%{?rhel} >= 9]
(License): Update the license expression and add an appropriate comment.
(%changelog): Mention it.