tests/sched_xetaffinity: do not assume that cpuset_size is big enough
Commit v5.15~160 ("tests/sched_xetaffinity: add a check
for a non-properly-sized CPU set") added a check that implicitly assumed
that cpuset_size is at least 8, which is not the case on 32-bit userspace
with less than 33 CPUs, and that has led to skipping the test altogether
on such systems due to syscall returning EFAULT.  Fix it by allocating
a separate set of the expected size and perform the check on it.

* tests/sched_xetaffinity.c (main): Add crop_size constant, set to 8;
add crop_cpuset variable;  use them instead of cpuset_size and cpuset,
respectively, in the non-properly-sized CPU set check.

Fixes: v5.15~160 "tests/sched_xetaffinity: add a check for a non-properly-sized CPU set"
Reported-by: Sven Schnelle <svens@linux.ibm.com>