Introduce --kill-on-exit option
This new option instructs the tracer to set PTRACE_O_EXITKILL option for
all tracees and not to detach them on cleanup so they will not be left
running after the tracer exit.

* src/exitkill.c: New file.
* src/exitkill.h: Likewise.
* src/Makefile.am (libstrace_a_SOURCES): Add exitkill.c and exitkill.h.
* src/strace.c: Include "exitkill.h".
(init): Handle --kill-on-exit option.
(cleanup): Do not detach tracees if PTRACE_O_EXITKILL is added
to ptrace_setoptions.
(usage): Document --kill-on-exit option.
* doc/strace.1.in: Likewise.
* tests/bexecve.test: Check that --kill-on-exit option works
with -b execve.
* tests/kill-on-exit.sh: New file.
* tests/kill-on-exit.test: New test.
* tests/options-syntax.test: Check invalid use of --kill-on-exit option.
* tests/Makefile.am (MISC_TESTS): Add kill-on-exit.test.
(EXTRA_DIST): Add kill-on-exit.sh.
* NEWS: Mention this change.