Set PTRACE_O_EXITKILL option along with PTRACE_O_TRACESECCOMP
When --seccomp-bpf is enabled and -p/--attach option is not used,
a seccomp filter is being assigned to all tracees.  In that case also set
PTRACE_O_EXITKILL option for all tracee processes and do not detach them
on cleanup so they will not be left running after the tracer exit.

* doc/strace.1.in: Document this.
* src/strace.c (init) <seccomp_filtering>: Add PTRACE_O_EXITKILL
to ptrace_setoptions when all conditions are met.
* tests/filter_seccomp-exitkill.test: New test.
* tests/Makefile.am (MISC_TESTS): Add filter_seccomp-exitkill.test.
* NEWS: Mention this change.