strace: add support for -u UID:GID that bypasses lookup
When building strace statically linked using the "--username" switch,
along with a user name, will lead to strange errors. The reason is that
the calls to getpwnam() and initgroups() will fail because glibc will
still try to dlopen() the libnss modules which then fails in unexpected
ways.

Extend the supported values to allow a classic UID:GID pair to be given
as numeric values, so that statically linked strace no longer relies on
dynamic name lookup.

Originally-by: Michael Vogt <mvo@ubuntu.com>
Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>
Co-authored-by: Dmitry V. Levin <ldv@strace.io>