prctl: organise the prctl option switch
It seems to be used to be kept in lexicographical order,
but over the time the new options were added to various places,
to the point it is not clear where to add it them now.   Let's put
groups of common option handlers at the top (and the one
that corresponds the default handling at the bottom), and sort the rest
in the option number order, so the new options can be added at the end
without a doubt.

* src/prctl.c (SYS_FUNC(prctl)) <switch>: Reorder.