tests: limit the number of binaries for the negative check in test_trace_expr
Part of the test in the test_trace_expr function is to check that only
the provided list of binaries produces the expected output, which
is implemented by checking for the absence of output against
(the rest of) the binaries provided in pure_executables.list.
As the number of tests implemented in the strace test suite grew
over time, the execution time of such tests grew as well, which makes
the test suite runs longer and tends to hit the (already extremely
generous) test execution time limit more often on underpowered
and/or heavily utilised machines.  Limit the test execution time
by randomly sampling the binaries to test for the absence of output.
This makes the check probabilistic, but the probability rather quickly
approaches 100% with the number of test runs strace undergoes during
each development cycle.

* tests/init.sh (test_trace_expr): Filter the result of the program set
subtraction through shuf -n before storing it in negative.list.
(TRACE_TESTS_SAMPLE): New environment variable to control the sample
size.