rtnl_addr: print the INFINITY_LIFE_TIME hint
The value 0xffffffff of ifa_prefered and ifa_valid fields of struct
ifa_cacheinfo is treated as infinite lifetime by the kernel,
and internally it even has the appropriate constant defined,
INFINITY_LIFE_TIME (in include/net/addrconf.h);  unfortunately,
this constant is not exposed to UAPI and cannot be freely used;
however, it can be printed as a comment to provide a hint
to the user that the relevant fields values are treated in a special way
(especially considering the fact that they are normally printed in
decimal and it is not obvious that the value 4294967295 might be treated
differently by the look of it).

* src/xlat/ifa_lifetimes.in: New file.
* src/rtnl_addr.c: Include "xlat/ifa_lifetimes.h".
(decode_ifa_cacheinfo): Use PRINT_FIELD_XVAL_U_VERBOSE
(with ifa_lifetimes as xlat) instead of PRINT_FIELD_U for printing
ifa_prefered and ifa_valid fields.
* tests/nlattr_ifaddrmsg.c [!XLAT_RAW] (main): Update the expected
output.