tests: Reduce expected precision for relative-timestamps tests
Currently we have a very small amount of tolerance around the
expected 1.0s sleep time, specifically 0.1s in either direction.

If the machine is not very fast (e.g. StarFive VisionFive 2, a
very popular and reasonably performant riscv64 board that's now
a couple of years old) and is heavily loaded (e.g. because other
builds are running at the same time or the test suite has been
executed with 'make -j check'), then the test might fail because
of an additional delay that's as small as 0.1s.

Since the test is really intended to ensure that the timestamps
are displayed using the requested precision, reporting a failure
in this case is not very useful. Instead, we should tolerate
delays of up to 1.0s as long as the timestamps are printed out
using the expected format.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>