tests/gen_tests.sh: use run_strace_match_diff for syscall tampering -y
As strace -y nees /proc filesystem, all tests that use strace -y have to
be skipped if /proc is unavailable.  This is usually done by the tests
themselves returning 77 if /proc is unavailable.  However, in the
generated tests that use both -y and syscall tampering the return code
was ignored.

* tests/gen_tests.sh: Generate run_strace_match_diff invocations
for the tests that use both -y and syscall tampering.