secontext: fix /proc/self handling
SELinux context queries are performed through prefixing by
/proc/<PID>/{root,cwd,fd/<FD>}/.  This led to querying strace's
namespace instead of PID's namespace when resolving paths such as
/proc/self/something, because /proc/self always points to ourselves
(hence strace).  Because of this /proc/self must be substituted as
/proc/<PID> first, but this may not always work, because there could
be a "fake" /proc/self" in the path.  To solve this, the context query
is performed on the substituted path and if there is no result, the
original path is queried as fallback.

Co-authored-by: Dmitry V. Levin <ldv@strace.io>