tests: fix uio.test with musl-libc
musl-libc handles the pwrite calls differently than glibc, making
the uio test fail.  Using glibc, there are separated pwrite and pwritev
calls, however musl uses only pwritev2 for both pwrite and pwritev.

To account to this difference, modify the expected matching regexp
to also match the musl output (append it with an OR).

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>