tests: use .code16 to express kvm input directly in real-mode
In the original proposal, the code passed to the virtual machine
was assumed being executed in real-mode:

https://lists.strace.io/pipermail/strace-devel/2017-December/007764.html

    +	const uint8_t code[] = {
    +		0xba, 0xf8, 0x03, /* mov $0x3f8, %dx */
    +		0x00, 0xd8,       /* add %bl, %al */
    +		0x04, '0',        /* add $'0', %al */
    ...

This code was derived from https://lwn.net/Articles/658511/.
code[] was well understandable.

However, when merging as 4f47c43ce73abf0398eb7f080a35fd3c5c1ff401,
the code was arranged like:

    +code(void)
    +{
    +   __asm__("mov $0xd80003f8, %edx; mov $'\n', %al; out %al, (%dx); hlt");

The notation was converted to protect-mode(e.g. %edx) though the gas
output of __asm__ (...) is the same as the code[] in the original
proposal.  I don't understand the intention and profit behind this
conversion.

The code was transformed again in the development history,
101621d4c7aac7be44570477480de04181236a09:

    +__asm__(
    +   ".type code, @object        \n"
    +   "code:              \n"
    +   "   mov $0xd80003f8, %edx   \n"

Just after making a virtual machine, it is in real-mode.  So writing
the code passed to the virtual machine in the real-mode in real-mode
notation is intuitive. It is hard to extend the the code passed to the
virtual machine in the real-mode in protect-mode notation.

* tests/ioctl_kvm_run_common.c (top-level): Use .code16 int
the inline assembly code. Rewrite the code in real-mode notation.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>